requirements.txt
django == 3.2.3
djangorestframework == 3.12.4
djangorestframework-simplejwt== 5.2.0
Pillow==8.0.0